requirements.txt
pandas 
numpy 
matplotlib 
seaborn 
jupyter 
pyyaml
xlrd
psycopg2-binary
python-dotenv